requirements.txt
# Core dependencies
pandas==2.1.4
numpy==1.24.4
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
python-dotenv==1.0.0

# Azure OpenAI
openai==1.3.8

# Data processing and analysis
scipy==1.11.4

# Utilities
tqdm==4.66.1
logging-config==1.0.3

# Development
pytest==7.4.3
black==23.11.0
flake8==6.1.0